+++ dacite/src/components/index.ts
@@ -15,3 +15,4 @@ export * from "./helpers";
 export * from "./unit";
 export * from "./section";
 export * from "./ssr-data-helpers";
+export * from "./unit-section-table";
