+++ dacite/src/components/unit/index.ts
@@ -3,4 +3,4 @@ export * from "./minimal-unit-editor";
 export * from "./color";
 export * from "./interval-suggest";
 export * from "./unit-editor";
-export * from "./table";
+export * from "../table";
